extract DocumentToReactComponents component
