+++ src/components/TemplateComponents/index.tsx
@@ -1,1 +1,2 @@
 export { default as Discussion } from "./Discussion"
+export { default as DocumentToReactComponents } from "./Contentful"
